[FEATURE] Rework history
Separate reccuring payments history from regular
